refactoring to class, add persist
