+++ src/config.lua
@@ -1,0 +1,30 @@
+local conf = {}
+
+local garbage = { "minecraft:water", "minecraft:lava", "minecraft:dirt", "minecraft:grass_block", "minecraft:stone",
+    "minecraft:cobblestone", "minecraft:diorite", "twigs:pebble", "minecraft:granite", "minecraft:gravel",
+    "minecraft:sand", "byg:soapstone", "minecraft:flint", "upgrade_aquatic:embedded_ammonite", "minecraft:torch",
+    "minecraft:deepslate", "minecraft:cobbled_deepslate", "twigs:rhyolite", "forbidden_arcanus:darkstone", "tetra:geode",
+    "silentgear:bort", "minecraft:tuff" }
+local fallingEntities = { "minecraft:sand", "minecraft:gravel" }
+local authorizedFuelSource = { "minecraft:coal" }
+
+-- Constant loading
+-- Falling entities
+conf.isFallingEntity = {}
+for _, v in ipairs(fallingEntities) do
+    conf.isFallingEntity[v] = true
+end
+-- fuel entities
+conf.isFuel = {}
+for _, v in ipairs(authorizedFuelSource) do
+    conf.isFuel[v] = true
+end
+-- trash
+conf.isTrash = {}
+for _, g in ipairs(garbage) do
+    conf.isTrash[g] = true
+end
+
+conf.FUEL_THRESHOLD = 10
+
+return conf
